+++ lib/index.ts
@@ -6,3 +6,4 @@
 export { download, downloadAndUnzipVSCode } from './download';
 export { runTests } from './runTest';
 export { resolveCliPathFromVSCodeExecutablePath, resolveCliArgsFromVSCodeExecutablePath } from './util';
+export * from './progress';